chore: added api to get operations in body and create transactions
This feature allows apps to send the parameters required to perform an operation for a user in the body alongwith the access token of the user and PeerID will create and broadcast the transaction all by itself. It removes the need for master password when creating a transaction for the user and also the need for creating the transaction in the frontend.